Fix NaN: load inference model in float32, remove bf16 cast
Model weights saved from bf16 training produced NaN logits when
cast again in serve.py. Load as float32 for stable inference.

+++ src/inference.py
@@ -40,7 +40,9 @@ class StyleClassifier:
         self.device = torch.device(device)
 
         self.tokenizer = AutoTokenizer.from_pretrained(model_path)
-        self.model = AutoModelForSequenceClassification.from_pretrained(model_path)
+        self.model = AutoModelForSequenceClassification.from_pretrained(
+            model_path, torch_dtype=torch.float32,
+        )
         self.model.to(self.device)
         self.model.eval()
